refactor: migrate User to Kotlin
