+++ src/main/java/com/likelionknu/applyserver/application/service/ApplicationQueryService.java
@@ -1,13 +1,13 @@
 package com.likelionknu.applyserver.application.service;
 
+import com.likelionknu.applyserver.auth.data.entity.User;
+import com.likelionknu.applyserver.auth.repository.UserRepository;
 import com.likelionknu.applyserver.application.data.dto.response.ApplicationDetailResponse;
 import com.likelionknu.applyserver.application.data.dto.response.ApplicationSummaryResponse;
 import com.likelionknu.applyserver.application.data.exception.ApplicationNotFoundException;
 import com.likelionknu.applyserver.application.data.exception.InvalidApplicationAccessException;
 import com.likelionknu.applyserver.application.data.exception.UserNotFoundException;
 import com.likelionknu.applyserver.application.data.repository.RecruitAnswerRepository;
-import com.likelionknu.applyserver.auth.data.entity.User;
-import com.likelionknu.applyserver.auth.data.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
