--- src/main/java/com/likelionknu/applyserver/auth/data/entity/User.java
@@ -1,61 +1,0 @@
-package com.likelionknu.applyserver.auth.data.entity;
-
-import com.likelionknu.applyserver.auth.data.enums.Role;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-@Table(name = "user")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(name = "profile_url", nullable = false)
-    private String profileUrl;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Role role;
-
-    @OneToOne(mappedBy = "user",
-            cascade = CascadeType.ALL,
-            orphanRemoval = true,
-            fetch = FetchType.LAZY)
-    private Profile profile;
-
-    // 사용자 삭제 시 함께 삭제되도록 하기 위한 연관관계
-    @OneToMany(mappedBy = "user",
-            cascade = CascadeType.ALL,
-            orphanRemoval = true)
-    private List<Application> applications = new ArrayList<>();
-
-    @UpdateTimestamp
-    @Column(name = "modified_at", nullable = false)
-    private LocalDateTime modifiedAt;
-
-    @Column(name = "last_access_at", nullable = false)
-    private LocalDateTime lastAccessAt;
-
-    @CreationTimestamp
-    @Column(name = "registered_at", nullable = false, updatable = false)
-    private LocalDateTime registeredAt;
-}
