+++ src/main/java/com/likelionknu/applyserver/auth/service/UserService.java
@@ -1,5 +1,7 @@
 package com.likelionknu.applyserver.auth.service;
 
+import com.likelionknu.applyserver.auth.data.entity.User;
+import com.likelionknu.applyserver.auth.repository.UserRepository;
 import com.likelionknu.applyserver.application.data.entity.Application;
 import com.likelionknu.applyserver.application.data.entity.MailHistory;
 import com.likelionknu.applyserver.application.data.entity.RecruitAnswer;
@@ -9,8 +11,6 @@ import com.likelionknu.applyserver.application.data.repository.RecruitAnswerRepo
 import com.likelionknu.applyserver.auth.data.dto.request.ModifyProfileRequestDto;
 import com.likelionknu.applyserver.auth.data.dto.response.ProfileResponseDto;
 import com.likelionknu.applyserver.auth.data.entity.Profile;
-import com.likelionknu.applyserver.auth.data.entity.User;
-import com.likelionknu.applyserver.auth.data.repository.UserRepository;
 import com.likelionknu.applyserver.auth.exception.UserNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
