+++ src/main/java/com/likelionknu/applyserver/recruit/service/RecruitService.java
@@ -1,10 +1,13 @@
 package com.likelionknu.applyserver.recruit.service;
 
+import com.likelionknu.applyserver.auth.data.entity.User;
+import com.likelionknu.applyserver.auth.repository.UserRepository;
+import com.likelionknu.applyserver.application.data.entity.Application;
+import com.likelionknu.applyserver.application.data.entity.RecruitAnswer;
+import com.likelionknu.applyserver.application.data.repository.ApplicationRepository;
 import com.likelionknu.applyserver.application.data.repository.RecruitAnswerRepository;
 import com.likelionknu.applyserver.auth.data.entity.Profile;
-import com.likelionknu.applyserver.auth.data.entity.User;
 import com.likelionknu.applyserver.auth.data.enums.ApplicationStatus;
-import com.likelionknu.applyserver.auth.data.repository.UserRepository;
 import com.likelionknu.applyserver.common.response.ErrorCode;
 import com.likelionknu.applyserver.common.response.GlobalException;
 import com.likelionknu.applyserver.common.security.SecurityUtil;
@@ -12,6 +15,8 @@ import com.likelionknu.applyserver.recruit.data.dto.response.RecruitAvailability
 import com.likelionknu.applyserver.recruit.data.dto.response.RecruitDetailResponse;
 import com.likelionknu.applyserver.recruit.data.dto.response.RecruitListResponse;
 import com.likelionknu.applyserver.recruit.data.dto.response.RecruitQuestionResponse;
+
+import com.likelionknu.applyserver.recruit.data.entity.Recruit;
 import com.likelionknu.applyserver.recruit.data.entity.RecruitContent;
 import com.likelionknu.applyserver.recruit.data.repository.RecruitContentRepository;
 import com.likelionknu.applyserver.recruit.data.repository.RecruitRepository;
@@ -113,7 +118,7 @@ public class RecruitService {
         log.info("[getRecruitQuestions] 공고 질문 조회: 공고 ID: {} 요청: {}", recruitId, user.getEmail());
 
         // 해당 공고에 대한 지원서 조회 (없을 수도 있음)
-        Optional<Application> applicationOpt = applicationRepository.findByUserIdAndRecruitId(user.getId(), recruitId);
+        Optional<Application> applicationOpt = Optional.ofNullable(applicationRepository.findByUserIdAndRecruitId(user.getId(), recruitId));
 
         // 답변 구성
         Map<Long, String> answerMap = new HashMap<>();
